feat(session): move prompt-audit ownership to SessionManager (#523)
Extracts prompt-audit logic from src/agents/acp/ into src/session/:

  src/session/audit-writer.ts   — file writer (moved from acp/prompt-audit.ts)
  src/session/audit.ts          — policy layer: enriches entries with stable
                                   sess-<uuid> identity before writing
  src/session/manager.ts        — auditPrompt() method + auditCallback injection
                                   in runInSession() so the adapter no longer
                                   decides audit policy

Key changes:
- SessionDescriptor gains projectDir for audit dir resolution
- AuditTurnEntry type (session/types.ts) carries raw adapter-reported data
- AgentRunOptions gains auditCallback? injected by SessionManager.runInSession
- ACP adapter run() path uses options.auditCallback (falls back to direct write
  when no session context)
- complete() path continues to call writePromptAudit directly (no session
  descriptor available — documented as scope limitation in spec)
- Audit files gain StableId header field (sess-<uuid>) for cross-hop correlation
- src/agents/acp/prompt-audit.ts becomes a re-export shim
- Tests migrated to test/unit/session/audit-writer.test.ts (+2 stableSessionId tests)
- mock-session-manager stub updated with auditPrompt no-op
- SPEC-session-manager-audit.md spec and ADR-011 updated

All 1211 tests pass, typecheck clean, lint clean.
