+++ src/session/index.ts
@@ -4,6 +4,7 @@
 
 export { SessionManager, _sessionManagerDeps } from "./manager";
 export type {
+  AuditTurnEntry,
   SessionDescriptor,
   SessionState,
   SessionRole,
@@ -13,3 +14,6 @@ export type {
   ISessionManager,
 } from "./types";
 export { SESSION_TRANSITIONS } from "./types";
+export { auditTurn } from "./audit";
+export { _promptAuditDeps, writePromptAudit, buildAuditFilename, findNaxProjectRoot } from "./audit-writer";
+export type { PromptAuditEntry } from "./audit-writer";
